[fix] domain 엔티티 정상적으로 스캔되도록 패키지 위치 변경
