+++ src/main/java/com/healthshop/healthshop/domain/item/Category.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain.item;
+package com.healthshop.healthshop.domain.item;
 
 import jakarta.persistence.*;
 import lombok.Getter;
