+++ src/main/java/com/healthshop/healthshop/domain/item/Item.java
@@ -1,10 +1,8 @@
-package com.healthshop.domain.item;
+package com.healthshop.healthshop.domain.item;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.List;
 
 @Entity
 @Getter @Setter
