+++ src/main/java/com/healthshop/healthshop/domain/member/cart/CartItem.java
@@ -1,29 +1,29 @@
-package com.healthshop.domain.order;
+package com.healthshop.healthshop.domain.member.cart;
 
-import com.healthshop.domain.item.Item;
+import com.healthshop.healthshop.domain.item.Item;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name = "order_item")
+@Table(name = "cart_item")
 @Getter @Setter
-public class OrderItem {
+public class CartItem {
 
     @Id @GeneratedValue
-    @Column(name = "order_item_id")
+    @Column(name = "cart_item_id")
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "order_id", nullable = false)
-    private Order order;
+    @JoinColumn(name = "cart_id", nullable = false)
+    private Cart cart;
 
     @ManyToOne
     @JoinColumn(name = "item_id", nullable = false)
     private Item item;
 
     @Column(nullable = false)
-    private Integer quantity;
+    private Integer quantity = 1;
 
     @Column(nullable = false)
     private Long price;
