+++ src/main/java/com/healthshop/healthshop/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain.order;
+package com.healthshop.healthshop.domain.order;
 
 public enum OrderStatus {
     COMPLETE, CANCEL
