+++ src/main/java/com/healthshop/healthshop/domain/order/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain.order;
+package com.healthshop.healthshop.domain.order;
 
 public enum PaymentMethod {
     CREDIT_CARD, ACCOUNT_TRANSFER
