+++ src/main/java/com/healthshop/healthshop/domain/order/delivery/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package com.healthshop.domain.order.delivery;
+package com.healthshop.healthshop.domain.order.delivery;
 
 public enum DeliveryStatus {
     BEFORE_SHIPPING, SHIPPING, COMPLETE
